refactor(chat): single chat-v2 tree — migrate shared components, delete legacy
chat-v2 (D-FE5) is the surviving chat tree. Migrate the components it
transitively depends on out of the legacy src/components/chat/ into
src/components/chat-v2/, repoint every importer, and delete the dead
remainder. src/components/chat/ is now gone.

Migrated into chat-v2 (git mv, history preserved): ChatMessage, ChatTimer,
SearchResults, reasoning-display, chain-of-thought, model-select,
free-models-banner, mini-chat-widget, guest-chat-counter (+ their 6 test
files). The plan's "6 files" undercounted the real closure: ChatMessage
dynamically imports SearchResults + reasoning-display, reasoning-display
imports chain-of-thought, and mini-chat-widget (landing TitleSection) had to
move too to clear the dir. `ai-sdk-chat-elements` in the plan's list does not
exist (only a .disabled file).

Deleted dead leaves: AudioPlayer, ai-sdk-model-option, performance-monitor,
ai-sdk-chat-elements.tsx.disabled, ai-sdk-chat.tsx.disabled (+ AudioPlayer test).

Repointed all @/components/chat/* imports to @/components/chat-v2/* across
app/chat, playground, inline-chat, TitleSection, features/chat/*, lib/hooks/*,
chat-ui-store, useRecentlyUsedModels, and external test suites (incl. dynamic
imports and jest.mock string paths).

Verification: grep "@/components/chat/" (non chat-v2) -> empty (gate passes);
lint clean; build green; full suite 21 failed/255 failed = baseline (zero new
failures; total drops only from the deleted dead AudioPlayer test). Dev-server
smoke: /chat 200, /models 200, /catalog/models 307->/models.
