+++ src/components/chat-v2/SearchResults.tsx
