+++ src/components/chat-v2/__tests__/free-models-banner.test.tsx
