+++ src/components/chat-v2/__tests__/mini-chat-widget.test.tsx
