+++ src/components/chat-v2/__tests__/model-select.test.tsx
