+++ src/components/chat-v2/__tests__/reasoning-display.test.tsx
