+++ src/components/chat-v2/chain-of-thought.tsx
