+++ src/components/chat-v2/guest-chat-counter.tsx
