+++ src/components/chat-v2/mini-chat-widget.tsx
