+++ src/components/chat-v2/reasoning-display.tsx
