--- src/components/chat/__tests__/AudioPlayer.test.tsx
@@ -1,443 +1,0 @@
-/**
- * Tests for AudioPlayer component
- */
-
-import React from 'react';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { AudioPlayer } from '../AudioPlayer';
-
-// Mock lucide-react icons
-jest.mock('lucide-react', () => ({
-  Play: () => <svg data-testid="play-icon" />,
-  Pause: () => <svg data-testid="pause-icon" />,
-  Volume2: () => <svg data-testid="volume-icon" />,
-  VolumeX: () => <svg data-testid="volume-x-icon" />,
-  Download: () => <svg data-testid="download-icon" />,
-  RotateCcw: () => <svg data-testid="restart-icon" />,
-}));
-
-// Mock ResizeObserver
-class MockResizeObserver {
-  observe = jest.fn();
-  unobserve = jest.fn();
-  disconnect = jest.fn();
-}
-
-global.ResizeObserver = MockResizeObserver as unknown as typeof ResizeObserver;
-
-// Mock play function that we can track and reset
-let mockPlay: jest.Mock;
-let mockPause: jest.Mock;
-
-// Mock HTMLMediaElement methods
-beforeAll(() => {
-  // Mock play to also dispatch the native 'play' event
-  mockPlay = jest.fn().mockImplementation(function(this: HTMLAudioElement) {
-    // Dispatch the native play event after a microtask
-    Promise.resolve().then(() => {
-      this.dispatchEvent(new Event('play'));
-    });
-    return Promise.resolve();
-  });
-
-  // Mock pause to also dispatch the native 'pause' event
-  mockPause = jest.fn().mockImplementation(function(this: HTMLAudioElement) {
-    this.dispatchEvent(new Event('pause'));
-  });
-
-  Object.defineProperty(HTMLMediaElement.prototype, 'play', {
-    configurable: true,
-    writable: true,
-    value: mockPlay,
-  });
-  Object.defineProperty(HTMLMediaElement.prototype, 'pause', {
-    configurable: true,
-    writable: true,
-    value: mockPause,
-  });
-  Object.defineProperty(HTMLMediaElement.prototype, 'load', {
-    configurable: true,
-    writable: true,
-    value: jest.fn(),
-  });
-});
-
-beforeEach(() => {
-  // Re-apply the mock implementations after clear to preserve event dispatching
-  mockPlay.mockClear();
-  mockPlay.mockImplementation(function(this: HTMLAudioElement) {
-    Promise.resolve().then(() => {
-      this.dispatchEvent(new Event('play'));
-    });
-    return Promise.resolve();
-  });
-
-  mockPause.mockClear();
-  mockPause.mockImplementation(function(this: HTMLAudioElement) {
-    this.dispatchEvent(new Event('pause'));
-  });
-});
-
-// Helper to simulate audio becoming ready (triggers canplay event)
-const simulateAudioReady = () => {
-  const audio = document.querySelector('audio');
-  if (audio) {
-    // Set duration property for proper time display
-    Object.defineProperty(audio, 'duration', { value: 120, writable: true });
-    audio.dispatchEvent(new Event('canplay'));
-    audio.dispatchEvent(new Event('loadedmetadata'));
-  }
-};
-
-describe('AudioPlayer', () => {
-  const mockSrc = 'data:audio/wav;base64,SGVsbG8gV29ybGQ=';
-
-  describe('Rendering', () => {
-    it('renders with default props', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Should have buttons (including play/pause)
-      const buttons = screen.getAllByRole('button');
-      expect(buttons.length).toBeGreaterThan(0);
-    });
-
-    it('renders with title', () => {
-      render(<AudioPlayer src={mockSrc} title="Test Audio" />);
-
-      expect(screen.getByText('Test Audio')).toBeInTheDocument();
-    });
-
-    it('renders in compact mode', () => {
-      const { container } = render(<AudioPlayer src={mockSrc} compact />);
-
-      // Compact mode has inline-flex class
-      expect(container.firstChild).toHaveClass('inline-flex');
-    });
-
-    it('renders with custom className', () => {
-      const { container } = render(
-        <AudioPlayer src={mockSrc} className="custom-class" />
-      );
-
-      expect(container.firstChild).toHaveClass('custom-class');
-    });
-  });
-
-  describe('Controls', () => {
-    it('shows play button initially', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Play icon should be visible (not pause)
-      expect(screen.getAllByRole('button').length).toBeGreaterThan(0);
-    });
-
-    it('has volume controls', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Should have volume/mute button
-      const muteButton = screen.getByTitle(/mute|unmute/i);
-      expect(muteButton).toBeInTheDocument();
-    });
-
-    it('has playback speed control', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Should have speed button showing "1x"
-      const speedButton = screen.getByTitle(/playback speed/i);
-      expect(speedButton).toBeInTheDocument();
-      expect(speedButton).toHaveTextContent('1x');
-    });
-
-    it('has download button', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      const downloadButton = screen.getByTitle(/download/i);
-      expect(downloadButton).toBeInTheDocument();
-    });
-
-    it('has restart button', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      const restartButton = screen.getByTitle(/restart/i);
-      expect(restartButton).toBeInTheDocument();
-    });
-  });
-
-  describe('Playback speed cycling', () => {
-    it('cycles through playback speeds', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      const speedButton = screen.getByTitle(/playback speed/i);
-
-      // Initial speed is 1x
-      expect(speedButton).toHaveTextContent('1x');
-
-      // Click to cycle
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('1.25x');
-
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('1.5x');
-
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('2x');
-
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('0.5x');
-
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('0.75x');
-
-      fireEvent.click(speedButton);
-      expect(speedButton).toHaveTextContent('1x'); // Back to start
-    });
-  });
-
-  describe('Callbacks', () => {
-    it('calls onPlay when play is triggered', async () => {
-      const onPlay = jest.fn();
-      render(<AudioPlayer src={mockSrc} onPlay={onPlay} />);
-
-      // Simulate audio becoming ready
-      simulateAudioReady();
-
-      // Find the main play button using data-testid
-      const playButton = screen.getByTestId('audio-play-pause-button');
-
-      // Wait for button to be enabled after canplay event
-      await waitFor(() => {
-        expect(playButton).not.toBeDisabled();
-      });
-
-      fireEvent.click(playButton);
-      // Wait for the async play() promise to resolve
-      await waitFor(() => {
-        expect(onPlay).toHaveBeenCalled();
-      });
-    });
-
-    it('calls onPause when pause is triggered', async () => {
-      const onPause = jest.fn();
-      render(<AudioPlayer src={mockSrc} onPause={onPause} />);
-
-      // Simulate audio becoming ready
-      simulateAudioReady();
-
-      // Find the main play button using data-testid
-      const playButton = screen.getByTestId('audio-play-pause-button');
-
-      // Wait for button to be enabled
-      await waitFor(() => {
-        expect(playButton).not.toBeDisabled();
-      });
-
-      fireEvent.click(playButton); // Play
-      // Wait for play to complete
-      await waitFor(() => {
-        expect(mockPlay).toHaveBeenCalled();
-      });
-      fireEvent.click(playButton); // Pause
-      expect(onPause).toHaveBeenCalled();
-    });
-  });
-
-  describe('Compact mode', () => {
-    it('shows minimal controls in compact mode', () => {
-      render(<AudioPlayer src={mockSrc} compact />);
-
-      // Compact mode should have play/pause button
-      expect(screen.getAllByRole('button').length).toBe(1);
-    });
-
-    it('shows time display in compact mode', () => {
-      render(<AudioPlayer src={mockSrc} compact />);
-
-      // Should show time like "0:00 / 0:00"
-      expect(screen.getByText(/0:00/)).toBeInTheDocument();
-    });
-  });
-
-  describe('Error handling', () => {
-    it('displays error when audio fails to load', async () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Trigger error event
-      const audio = document.querySelector('audio');
-      if (audio) {
-        audio.dispatchEvent(new Event('error'));
-      }
-
-      // Should show error message
-      await waitFor(() => {
-        expect(screen.getByText(/failed to load audio/i)).toBeInTheDocument();
-      });
-    });
-
-    it('handles play failure gracefully', async () => {
-      // Mock play to reject
-      mockPlay.mockRejectedValueOnce(new Error('Autoplay blocked'));
-
-      const onPlay = jest.fn();
-      render(<AudioPlayer src={mockSrc} onPlay={onPlay} />);
-
-      // Simulate audio becoming ready
-      simulateAudioReady();
-
-      const playButton = screen.getByTestId('audio-play-pause-button');
-
-      await waitFor(() => {
-        expect(playButton).not.toBeDisabled();
-      });
-
-      fireEvent.click(playButton);
-
-      // Should show error and NOT call onPlay
-      await waitFor(() => {
-        expect(screen.getByText(/failed to play audio/i)).toBeInTheDocument();
-      });
-      expect(onPlay).not.toHaveBeenCalled();
-    });
-
-    it('handles autoplay failure in handleCanPlay', async () => {
-      // Mock play to reject for autoplay
-      mockPlay.mockRejectedValueOnce(new Error('Autoplay blocked'));
-
-      const onPlay = jest.fn();
-      render(<AudioPlayer src={mockSrc} autoPlay onPlay={onPlay} />);
-
-      // Trigger canplay event (which triggers autoplay)
-      const audio = document.querySelector('audio');
-      if (audio) {
-        Object.defineProperty(audio, 'duration', { value: 120, writable: true });
-        audio.dispatchEvent(new Event('canplay'));
-      }
-
-      // Should show error and NOT call onPlay
-      await waitFor(() => {
-        expect(screen.getByText(/failed to play audio/i)).toBeInTheDocument();
-      });
-      expect(onPlay).not.toHaveBeenCalled();
-    });
-
-    it('handles restart play failure', async () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Simulate audio becoming ready
-      simulateAudioReady();
-
-      // Mock play to reject for restart
-      mockPlay.mockRejectedValueOnce(new Error('Play blocked'));
-
-      const restartButton = screen.getByTestId('audio-restart-button');
-      fireEvent.click(restartButton);
-
-      // Should show error
-      await waitFor(() => {
-        expect(screen.getByText(/failed to play audio/i)).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Time formatting', () => {
-    it('formats time correctly', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Initial time should show 0:00
-      const timeDisplays = screen.getAllByText(/0:00/);
-      expect(timeDisplays.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('Source changes', () => {
-    it('resets state when src changes', async () => {
-      const { rerender } = render(<AudioPlayer src={mockSrc} />);
-
-      // Simulate audio becoming ready
-      simulateAudioReady();
-
-      await waitFor(() => {
-        const playButton = screen.getByTestId('audio-play-pause-button');
-        expect(playButton).not.toBeDisabled();
-      });
-
-      // Play the audio
-      const playButton = screen.getByTestId('audio-play-pause-button');
-      fireEvent.click(playButton);
-      await waitFor(() => {
-        expect(mockPlay).toHaveBeenCalled();
-      });
-
-      // Change src
-      const newSrc = 'data:audio/wav;base64,TmV3QXVkaW8=';
-      rerender(<AudioPlayer src={newSrc} />);
-
-      // State should be reset - isLoading should be true again (button disabled)
-      await waitFor(() => {
-        const button = screen.getByTestId('audio-play-pause-button');
-        expect(button).toBeDisabled();
-      });
-    });
-
-    it('calls audio.load() when src changes', async () => {
-      const mockLoad = jest.fn();
-      Object.defineProperty(HTMLMediaElement.prototype, 'load', {
-        configurable: true,
-        writable: true,
-        value: mockLoad,
-      });
-
-      const { rerender } = render(<AudioPlayer src={mockSrc} />);
-      mockLoad.mockClear(); // Clear initial load call
-
-      // Change src
-      const newSrc = 'data:audio/wav;base64,TmV3QXVkaW8=';
-      rerender(<AudioPlayer src={newSrc} />);
-
-      // audio.load() should be called
-      expect(mockLoad).toHaveBeenCalled();
-    });
-  });
-
-  describe('Volume controls', () => {
-    it('unmutes audio when volume is adjusted above 0', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      const audio = document.querySelector('audio');
-      if (audio) {
-        // First mute the audio
-        audio.muted = true;
-      }
-
-      // Find and interact with volume slider (implementation depends on your Slider component)
-      // This tests that audio.muted is set to false when volume > 0
-      expect(audio?.muted).toBeDefined();
-    });
-  });
-
-  describe('Accessibility', () => {
-    it('has accessible buttons when audio is ready', async () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      // Simulate audio becoming ready (play button is disabled while loading)
-      simulateAudioReady();
-
-      // Wait for React to process the event and update state
-      await waitFor(() => {
-        const buttons = screen.getAllByRole('button');
-        buttons.forEach(button => {
-          expect(button).not.toBeDisabled();
-        });
-      });
-    });
-
-    it('has title attributes for icon buttons', () => {
-      render(<AudioPlayer src={mockSrc} />);
-
-      expect(screen.getByTestId('audio-restart-button')).toHaveAttribute('title', 'Restart');
-      expect(screen.getByTestId('audio-speed-button')).toHaveAttribute('title', 'Playback speed');
-      expect(screen.getByTestId('audio-mute-button')).toHaveAttribute('title');
-      expect(screen.getByTestId('audio-download-button')).toHaveAttribute('title', 'Download audio');
-    });
-  });
-});
